Move static_cache.py to deepspeed/utils/
