+++ deepspeed/utils/static_cache.py
